Refactored into classes and adding stopping rules
